pyproject.toml
[project]
name = "arca-delineo"
version = "0.1.0"
description = "GenAI pipeline to transform UI hand sketches with prompt in high level mockups via open-source diffusion model's control techniques."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "albumentations>=2.0.8",
    "gradio>=6.0.1",
    "ipykernel>=7.1.0",
    "logging>=0.4.9.6",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "omegaconf>=2.3.0",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=12.0.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
]
